Exchanged the file names of DocMyAppointment & DocPatientRecord. The button works. Making necessary Changes now.
